requirements.txt
flask
flask-cors
azure-ai-formrecognizer
gunicorn